Address comments, introduce task interface
