+++ pkg/task/edit_test.go
@@ -86,7 +86,7 @@ info:
 		repoOpener:        repoOpener,
 	}
 
-	res, _, err := epm.Apply(context.Background(), repository.PackageResources{})
+	res, _, err := epm.apply(context.Background(), repository.PackageResources{})
 	if err != nil {
 		t.Errorf("task apply failed: %v", err)
 	}
